perf(dashboard): Improve efficiency of npm plugin scan
